pyproject.toml
[tool.black]
line-length = 110
target-version = ["py310"]

[tool.isort]
line_length = 110
include_trailing_comma = true
known_local_folder = ["src", "tests", "transactions_api"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "transactions_api.settings"
addopts = "-vv"

[tool.poetry]
name = "transactions-api"
version = "0.1.0"
description = "Rest API to deal with simple transactions"
authors = ["Vinicius Chan <vchanbr@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "==4.0.3"
djangorestframework = "==3.13.1"
django-rest-swagger = "*"
python-decouple = "*"
psycopg2 = "*"
dj-database-url = "*"


[tool.poetry.dev-dependencies]
ipython = "*"
factory-boy = "*"
faker = "*"
pre-commit = "*"
pytest = "*"
pytest-django = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
